requirements.txt
fastapi==0.115.7
openai_whisper==20240930
Requests==2.32.3
streamlit==1.41.1
torch==2.5.1
torchaudio==2.5.1
uvicorn
python-multipart
